# Please enter a commit message to explain why this merge is necessary,
# especially if it merges an updated upstream into a topic branch.
#
# Lines starting with '#' will be ignored, and an empty message aborts
# the commit.

+++ web/src/main/webapp/WEB-INF/jsp/submit/uploadFileToMirth.jsp
@@ -166,15 +166,26 @@
                             </td>
                             
                             <td width=12.5%  align="center">
+							    <%
+								long passTime = System.currentTimeMillis() - fileObjects[i].lastModified();
+                                
+   						        if(passTime >= 5000){
+									
+								%>
                                 <a target="_new" href="UploadCRFData?action=download&fileId=<%=fileNames[i]%>">
 									<span name="bt_Download1" class="icon icon-download" border="0" align="left" hspace="6"
 										 alt="<fmt:message key="download" bundle="${resword}"/>" title="<fmt:message key="download" bundle="${resword}"/>">
-								</a>
+								</a>								
 								<a href="UploadCRFData?action=delete&fileId=<%=fileNames[i]%>">
 									<span name="bt_Delete1" class="icon icon-trash red" border="0" alt="<fmt:message key="delete" bundle="${resword}"/>"
 										 title="<fmt:message key="delete" bundle="${resword}"/>" align="left" hspace="6"
 										 onClick='return confirm("Please confirm that you want to delete log file <%=fileNames[i]%>");'>
 								</a>
+								<%
+								} else{
+								%>
+								processing...
+								<%}%>
                             </td>
                         </tr>
